refactor(api): fold files/by-path into the canonical surface (#613)
